消息队列 添加商品后更新索引库

+++ dd-manager/dd-manager-interface/src/main/java/com/lbf/ddshop/service/ItemService.java
@@ -29,5 +29,5 @@ public interface ItemService {
 
     int up_updateItemsByIds(List<Long> ids);
 
-    int saveItem(TbItem tbItem, String content,String paramData);
+    Long saveItem(TbItem tbItem, String content,String paramData);
 }
